Revert "Test implementation"
This reverts commit 14c979d9543d9ca3b9a37e52713ef0dbbd230023.
