--- source/Kernel-NativeAOT/10-System/Plugs/Console.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace System
-{
-    public class Console
-    {
-        public static void WriteLine(string str)
-        {
-            Loader.Program.Write(str);
-        }
-    }
-}
