--- source/Kernel-NativeAOT/10-System/Runtime/CompilerHelpers.cs
@@ -1,20 +1,0 @@
-﻿using System;
-
-namespace Internal.Runtime.CompilerHelpers
-{
-    using System.Runtime;
-
-    class StartupCodeHelpers
-    {
-        [RuntimeExport("RhpReversePInvoke2")]
-        static void RhpReversePInvoke2() { }
-        [RuntimeExport("RhpReversePInvokeReturn2")]
-        static void RhpReversePInvokeReturn2() { }
-        [System.Runtime.RuntimeExport("__fail_fast")]
-        static void FailFast() { while (true) ; }
-        [System.Runtime.RuntimeExport("RhpPInvoke")]
-        static void RphPinvoke() { }
-        [System.Runtime.RuntimeExport("RhpPInvokeReturn")]
-        static void RphPinvokeReturn() { }
-    }
-}
