--- source/Kernel-NativeAOT/10-System/Runtime/InteropServices.cs
@@ -1,28 +1,0 @@
-﻿using System;
-
-namespace System.Runtime.InteropServices
-{
-    public class UnmanagedType { }
-
-    sealed class StructLayoutAttribute : Attribute
-    {
-        public StructLayoutAttribute(LayoutKind layoutKind)
-        {
-        }
-    }
-
-    internal enum LayoutKind
-    {
-        Sequential = 0, // 0x00000008,
-        Explicit = 2, // 0x00000010,
-        Auto = 3, // 0x00000000,
-    }
-
-    internal enum CharSet
-    {
-        None = 1,       // User didn't specify how to marshal strings.
-        Ansi = 2,       // Strings should be marshalled as ANSI 1 byte chars.
-        Unicode = 3,    // Strings should be marshalled as Unicode 2 byte chars.
-        Auto = 4,       // Marshal Strings in the right way for the target system.
-    }
-}
