--- source/Kernel-NativeAOT/10-System/Runtime/RuntimeExportAttribute.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace System.Runtime
-{
-    internal sealed class RuntimeExportAttribute : Attribute
-    {
-        public RuntimeExportAttribute(string entry) { }
-    }
-}
