--- source/Kernel-NativeAOT/10-System/Runtime/UnmanagedCallersOnlyAttribute.cs
@@ -1,21 +1,0 @@
-﻿
-namespace System.Runtime.InteropServices
-{
-
-    public enum CallingConvention
-    {
-        Winapi = 1,
-        Cdecl = 2,
-        StdCall = 3,
-        ThisCall = 4,
-        FastCall = 5,
-    }
-
-    public sealed class UnmanagedCallersOnlyAttribute : Attribute
-    {
-        public string EntryPoint;
-        public CallingConvention CallingConvention;
-
-        public UnmanagedCallersOnlyAttribute() { }
-    }
-}
