--- source/Kernel-NativeAOT/10-System/Types.cs
@@ -1,51 +1,0 @@
-﻿using System;
-using System.Runtime.InteropServices;
-
-namespace System
-{
-    public class Object { public IntPtr m_pEEType; } // The layout of object is a contract with the compiler.
-    public struct Void { }
-    public struct Boolean { }
-    public struct Char { }
-    public struct SByte { }
-    public struct Byte { }
-    public struct Int16 { }
-    public struct UInt16 { }
-    public struct Int32 { }
-    public struct UInt32 { }
-    public struct Int64 { }
-    public struct UInt64 { }
-    public struct IntPtr { }
-    public struct UIntPtr { }
-    public struct Single { }
-    public struct Double { }
-    public abstract class ValueType { }
-    public abstract class Enum : ValueType { }
-    public struct Nullable<T> where T : struct { }
-
-    public abstract class String { public readonly int Length; }
-    public abstract class Array { }
-    public abstract class Delegate { }
-    public abstract class MulticastDelegate : Delegate { }
-
-    public struct RuntimeTypeHandle { }
-    public struct RuntimeMethodHandle { }
-    public struct RuntimeFieldHandle { }
-
-    public class Array<T> : Array { }
-
-    public class Attribute { }
-
-    namespace Runtime.CompilerServices
-    {
-        public class RuntimeHelpers
-        {
-            public static unsafe int OffsetToStringData => sizeof(IntPtr) + sizeof(int);
-        }
-
-        public static class RuntimeFeature
-        {
-            public const string UnmanagedSignatureCallingConvention = nameof(UnmanagedSignatureCallingConvention);
-        }
-    }
-}
