--- source/Kernel-NativeAOT/30-HAL/Boot.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace Cosmos.HAL
-{
-    public class Boot
-    {
-        public static void Init()
-        {
-            Console.WriteLine("Cosmos.HAL booted successfully.");
-            System.Boot.Init();
-        }
-    }
-}
