--- source/Kernel-NativeAOT/40-System/Boot.cs
@@ -1,20 +1,0 @@
-﻿using System;
-
-namespace Cosmos.System
-{
-    public class Boot
-    {
-        public static void Init()
-        {
-            Console.WriteLine("Cosmos.System booted successfully.");
-
-            var kernel = new Kernel();
-            kernel.BeforeRun();
-
-            while (true)
-            {
-                kernel.Run();
-            }
-        }
-    }
-}
