--- source/Kernel-NativeAOT/40-System/Kernel.cs
@@ -1,17 +1,0 @@
-﻿using System;
-
-namespace Cosmos.System
-{
-    public class Kernel
-    {
-        public void BeforeRun()
-        {
-            Console.WriteLine("Kernel booted successfully.");
-        }
-
-        public void Run()
-        {
-
-        }
-    }
-}
