--- source/Kernel-NativeAOT/Cosmos.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31129.286
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cosmos", "Cosmos.csproj", "{848F6F02-FD0E-4F01-8134-5A34C46E92DE}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{848F6F02-FD0E-4F01-8134-5A34C46E92DE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{848F6F02-FD0E-4F01-8134-5A34C46E92DE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{848F6F02-FD0E-4F01-8134-5A34C46E92DE}.Debug|x64.ActiveCfg = Debug|x64
-		{848F6F02-FD0E-4F01-8134-5A34C46E92DE}.Debug|x64.Build.0 = Debug|x64
-		{848F6F02-FD0E-4F01-8134-5A34C46E92DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{848F6F02-FD0E-4F01-8134-5A34C46E92DE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{848F6F02-FD0E-4F01-8134-5A34C46E92DE}.Release|x64.ActiveCfg = Release|x64
-		{848F6F02-FD0E-4F01-8134-5A34C46E92DE}.Release|x64.Build.0 = Release|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F5CB813B-8BF7-4847-964A-09DA7B1CE53F}
-	EndGlobalSection
-EndGlobal
